+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is (loosely) based on
 project more or less adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## Release 0.2.10
+
+2022-03-01 22:20 CET
 
 * Support `!embed` blocks for youtube, in preparation for improving
   privacy by not contacting youtube before the user starts a video.